refactor(Realtime): Simplify
