+++ packages/cozy-realtime/src/Socket.js
@@ -3,28 +3,10 @@ import MicroEE from 'microee'
 import pickBy from 'lodash/pickBy'
 
 /**
- * Return websocket url from cozyClient
+ * Socket class
  *
- * @return {String}  WebSocket url
+ * @class
  */
-export const getWebSocketUrl = cozyClient => {
-  const isSecureURL = url => !!url.match(`^(https:/{2})`)
-
-  const url = cozyClient.stackClient.uri
-  const protocol = isSecureURL(url) ? 'wss:' : 'ws:'
-  const host = new URL(url).host
-
-  return `${protocol}//${host}/realtime/`
-}
-
-/**
- * Return token from cozyClient
- *
- * @return {String}  token
- */
-export const getWebSocketToken = cozyClient =>
-  cozyClient.stackClient.token.accessToken || cozyClient.stackClient.token.token
-
 class Socket {
   /**
    * Doctype name on cozy stack
@@ -68,16 +50,13 @@ class Socket {
     this._token = token
   }
 
-  isConnected() {
+  isOpen() {
     return !!(this._socket && this._socket.readyState === WebSocket.OPEN)
   }
 
   updateAuthentication(token) {
     this._token = token
-
-    if (this.isConnected()) {
-      this.authentication()
-    }
+    this.authentication()
   }
 
   /**
@@ -88,17 +67,26 @@ class Socket {
   connect() {
     return new Promise((resolve, reject) => {
       this._socket = new WebSocket(this._url, this._doctype)
-      const emit = type => event => this.emit(type, event)
 
-      this._socket.onmessage = emit('onmessage')
-      this._socket.onclose = emit('onclose')
+      this._socket.onmessage = event => {
+        const data = JSON.parse(event.data)
+        const eventName = data.event.toLowerCase()
+        const { type, id, doc } = data.payload
+
+        this.emit('message', { type, id, eventName }, doc)
+      }
+
+      this._socket.onclose = event => this.emit('close', event)
+
       this._socket.onerror = error => {
         this._socket = null
-        this.emit('onerror', error)
+        this.emit('error', error)
         reject(error)
       }
+
       this._socket.onopen = event => {
-        this.emit('onopen', event)
+        this.authentication()
+        this.emit('open', event)
         resolve(event)
       }
     })
@@ -110,7 +98,7 @@ class Socket {
    * @see https://github.com/cozy/cozy-stack/blob/master/docs/realtime.md#auth
    */
   authentication() {
-    if (this.isConnected()) {
+    if (this.isOpen()) {
       this._socket.send(
         JSON.stringify({ method: 'AUTH', payload: this._token })
       )
@@ -126,10 +114,9 @@ class Socket {
    * @param {String} type  Document doctype to subscribe to
    * @param {String} id  Document id to subscribe to (not required)
    */
-  async subscribe(type, id = null) {
-    if (!this.isConnected()) {
+  async subscribe(type, id = undefined) {
+    if (!this.isOpen()) {
       await this.connect()
-      this.authentication()
     }
 
     const payload = pickBy({ type, id })
@@ -143,11 +130,14 @@ class Socket {
     this.emit('subscribe')
   }
 
+  /**
+   * Close socket
+   */
   close() {
-    if (this.isConnected()) {
+    if (this.isOpen()) {
       this._socket.close()
-      this._socket = null
     }
+    this._socket = null
   }
 }
